Center of pressure of each blade element.
This is another input file for the MATLAB source script. These data are
the output of rotating reference model implemented in RANS-solver (e.g.
SRF in ANSYS FLUENT). These data are the two coordinates of the center
of pressure for each blade section. They are used in the MATLAB script
to locate the line of action of force components on each blade section.